Moved settings files and added hover effect to module item
